chore: address issue #409
